LayoutTests/media/modern-media-controls/scrubber/scrubber-has-correct-ax-label.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<style type="text/css" media="screen">

    video, #host {
        position: absolute;
        top: 0;
        left: 0;
        width: 800px;
        height: 240px;
    }

</style>
<video src="../../content/test.mp4" style="width: 320px; height: 240px;" controls></video>
<div id="host"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing for an accessibility text on the scrubber that correctly reflects the changes in media.currentTime.");

const container = document.querySelector("div#host");
const media = document.querySelector("video");
const mediaController = createControls(container, media, null);

mediaController.controls.showsStartButton = false;

const scrubberNodes = mediaController.controls.timeControl.scrubber.children;
const scrubberInput = scrubberNodes[scrubberNodes.length - 1].element;

media.addEventListener("canplay", () => {

    shouldBeEqualToString("scrubberInput.getAttribute('aria-valuetext')", "0 Minute 0 Second");
    media.currentTime = 2;
    shouldBe("media.currentTime", "2");
    shouldBecomeEqual("scrubberInput.getAttribute('aria-valuetext') === '0 Minute 2 Seconds'", "true", finishJSTest);
})

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
